+++ api_v3/lib/types/filters/ValidateAccessResponseProfile.php
@@ -19,6 +19,8 @@ class ValidateAccessResponseProfile
 	 */
 	public static function validateAccess($relatedFilter)
 	{
+		if (self::PartnerExcludedFromResponseProfile())
+			return true;
 		$clazz = self::getServiceClassInstance($relatedFilter);
 		return self::validate($clazz);
 	}
@@ -86,6 +88,17 @@ class ValidateAccessResponseProfile
 			throw  $e;
 		}
 		return true;
+	}
+
+	/**
+	 * @return bool
+	 * @throws Exception
+	 */
+	private static function PartnerExcludedFromResponseProfile()
+	{
+		$partnerId = kCurrentContext::getCurrentPartnerId();
+		$partnersToSkip = kConf::get('skip_response_profile_validation_partners','local', array());
+		return in_array($partnerId,$partnersToSkip);
 	}
 
 }