+++ release-notes.md
@@ -1,3 +1,16 @@
+# Naos 14.4.0 #
+
+## create CAPTIONS_EDITOR_ROLE for partner 0 ##
+- Issue Type: Task
+- Issue ID: KMS-18227
+
+### Configuration ###
+	None.
+
+#### Deployment Scripts ####
+
+	Run 'php /opt/kaltura/app/deployment/updates/scripts/2018_08_06_create_captions_editor_application_role.php'
+
 # Naos 14.3.0 #
 
 ## Allow media-server partner to get reports##